+++ packages/plugin-gallery/web/src/components/gallery-settings-modal.js
@@ -28,6 +28,8 @@ class ManageMediaSection extends Component {
     store.set('componentData', componentData);
   };
 
+  static contextType = GlobalContext;
+
   handleFileChange = (files, itemPos) => {
     if (files.length > 0) {
       const handleFilesSelected = this.props.store.getBlockHandler('handleFilesSelected');
@@ -41,18 +43,10 @@ class ManageMediaSection extends Component {
   };
 
   render() {
-    const {
-      helpers,
-      store,
-      t,
-      relValue,
-      anchorTarget,
-      isMobile,
-      uiSettings,
-      languageDir,
-      accept,
-    } = this.props;
+    const { helpers, store, t, relValue, anchorTarget, isMobile, uiSettings, accept } = this.props;
     const { handleFileSelection } = helpers;
+    const { languageDir } = this.context;
+
     return (
       <div dir={languageDir} className={styles.gallerySettings_tab_section}>
         <SortableComponent
@@ -137,6 +131,7 @@ class AdvancedSettingsSection extends Component {
             store={store}
             t={t}
             isMobile={isMobile}
+            languageDir={languageDir}
           />
         </div>
       )
@@ -288,6 +283,7 @@ export class GallerySettingsModal extends Component {
           store={this.props.pubsub.store}
           helpers={this.props.helpers}
           t={this.props.t}
+          languageDir={this.props.languageDir}
         />
       </Tab>
     ),